% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{mi_analyze_document}
\alias{mi_analyze_document}
\title{Run a one or more documents through a misinformation/bias sentiment analysis}
\usage{
mi_analyze_document(path = ".", pattern = NULL, bias_type, sentiment_list)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to
the working directory, \code{getwd()}. Tilde expansion (see `path.expand``) is
performed. Missing values will be ignored.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will be returned. (e.g. "\code{.*\\.txt$})}

\item{bias_type}{a one-word (no punctuation or spaces) description of the
bias type of the sentiment list being used. This will be used to tag
the output so it can be used later when aggregating results of multiple
bias-type analyses.}

\item{sentiment_list}{a character vector of words/phrases. The built-in sentiment
bias lists can be used via \code{\link[=mi_use_builtin]{mi_use_builtin()}} \emph{or} the caller can pass
in a charcter vector of lower-case words/phrases}
}
\value{
a \code{list} (classed as \code{misinfo_corpus}) with:
\itemize{
\item \code{bias_type} (the bias type label passed in to the function)
\item \code{corpus} (the raw data from the processed document corpus)
\item \code{frequency_summary} (frequncy summary for each document in corpus)
\item \code{individual_frequency} (individual component frequency for each phrase in
each document in the corpus)
}
}
\description{
Run a one or more documents through a misinformation/bias sentiment analysis
}
\examples{
mi_analyze_document(
  system.file("extdat", package="misinfo"),
  pattern=".*txt$",
  bias_type = "explanatory",
  sentiment_list = mi_use_builtin("explanatory")
)
}
