% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{mi_plot_document_summary}
\alias{mi_plot_document_summary}
\title{Plot raw frequency count summary from a processed corpus}
\usage{
mi_plot_document_summary(corpus, normalized = TRUE)
}
\arguments{
\item{corpus}{a document corpus processed with \code{\link[=mi_analyze_document]{mi_analyze_document()}}}

\item{normalized}{if \code{TRUE}, normalized values will be used, otherwise
raw counts will be used; Default : \code{TRUE}}
}
\value{
ggplot2 object
}
\description{
Plot raw frequency count summary from a processed corpus
}
\examples{
\dontrun{
mi_analyze_document(
  system.file("extdat", package="misinfo"),
  pattern=".*txt$",
  bias_type = "explanatory",
  sentiment_list = mi_use_builtin("explanatory")
) -> corpus

mi_plot_document_summary(corpus)
mi_plot_document_summary(corpus, FALSE)
}
}
