% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin-lists.R
\name{mi_use_builtin}
\alias{mi_use_builtin}
\title{Use a built-in 'sentiment' word/phrase list}
\usage{
mi_use_builtin(type = c("uncertainty", "sourcing", "retractors",
  "explanatory"))
}
\arguments{
\item{list_name}{name of built-in 'sentiment' word/phrase list}
}
\description{
The package comes with a set of built-in 'sentiment' word/phrase lists
for different classifications of "misinformation":\cr
\itemize{
\item \code{explanatory}
\item \code{retractors}
\item \code{sourcing}
\item \code{uncertainty}
Call the function with one of those names and the corresponding character
vector will be returned.
}
}
\examples{
mi_use_builtin("retractors")
}
