% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misinfo-package.R
\docType{package}
\name{misinfo}
\alias{misinfo}
\alias{misinfo-package}
\title{Tools to Perform 'Misinformation' Analysis on a Text Corpus}
\description{
Tools to Perform 'Misinformation' Analysis on a Text Corpus
}
\author{
Bob Rudis (bob@rud.is)
}
