% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-to-txt.R
\name{word_doc_to_txt}
\alias{word_doc_to_txt}
\title{Convert a Word document to a plaintext document}
\usage{
word_doc_to_txt(in_path, out_path)
}
\arguments{
\item{in_path, out_path}{source and destination files. These can be
abbreviated path names (e.g. \code{~/in/word.docx}, \code{~/out/word.txt}) but
they \emph{must} be paths to file names. Extensions are not automatically
added by this function}
}
\value{
the \code{path.expand}ed \code{out_path} (invisibly)
}
\description{
Convert a Word document to a plaintext document
}
\note{
Requires a working \code{pandoc} installation; this comes with RStudio, so
if the caller is using RStudio, \code{pandoc} should be available.
}
\examples{
\dontrun{
word_doc_to_txt("~/in/word.docx", "~/out/word.txt")
}
}
